Renamed mpdist.py
